backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
psycopg2-binary==2.9.9
black==24.8
ipython==8.27.0
